+++ CentroLink/CentroLink/PromoTicketSetupModule/ServicesData/PromoTicketSetupDataService.cs
@@ -87,5 +87,21 @@ namespace CentroLink.PromoTicketSetupModule.ServicesData
 
             return offSet;
         }
+
+        public int GetPrintPromo()
+        {
+            const string sql = @";EXEC [tpcGetPrintPromo]";
+            var printPromo = Db.ExecuteScalar<int>(sql);
+            return printPromo;
+        }
+        public void SetPrintPromo(bool printEntryTicket)
+        {
+            const string sql = @";EXEC [tpcSetPrintPromo] @PrintPromoValue";
+
+            Db.Execute(sql, new
+            {
+                PrintPromoValue = printEntryTicket
+            });
+        }
     }
 }